dev2 EzNoteDocumentListener merged with EzNoteDocument for simplification process
