+++ src/gdr/eznote/document/EzNoteDocument.java
@@ -1,6 +1,7 @@
 package gdr.eznote.document;
 
 import gdr.eznote.EzNoteFrame;
+import gdr.eznote.frames.*;
 import gdr.eznote.util.*;
 import java.io.File;
 import java.io.FileNotFoundException;
@@ -10,48 +11,46 @@ import java.util.NoSuchElementException;
 import java.util.Scanner;
 import javax.swing.JFileChooser;
 import javax.swing.JOptionPane;
+import javax.swing.event.DocumentEvent;
+import javax.swing.event.DocumentListener;
 import javax.swing.filechooser.FileNameExtensionFilter;
 
-public class EzNoteDocument {
+public class EzNoteDocument implements DocumentListener {
     
-    private EzNoteDocumentListener dc;
     private EzNoteFrameUtil util;
     private EzNoteFrame parent;
+    private int changeCounter;
+    private boolean changeIndicator;
     
     private File file;
     
     public EzNoteDocument(EzNoteFrame parent){
-        //init event listener for this document
-        this.dc = new EzNoteDocumentListener(parent);
         //init frame utilities for this document to handler display of some things
         this.util = new EzNoteFrameUtil(parent);
         //sets the documentlistener of the frame handled automatically
-        parent.setDocumentListener(dc);
+        parent.setDocumentListener(this);
         this.parent = parent;
     }
     
-    public EzNoteDocumentListener getDocumentListener(){
-        return this.dc;
+    public EzNoteFrameUtil getUtilities(){
+        return this.util;
     }
     
     public File getFile(){
         return this.file;
     }
     
-    public boolean open(){
+    public boolean open(File f, EzNoteFileChooser fc, int state){
         boolean complete = false;
         FileNameExtensionFilter filter = new FileNameExtensionFilter("Text Files", "txt", "text");
         this.parent.getFileChooser().setFileFilter(filter);
-        this.parent.getFileChooser().displayOpen(parent);
-        File f = this.parent.getFileChooser().getSelectedFile();
-        this.parent.setTitle(util.getWindowTitle(true, f.getName()));
-        if (f.canRead()) {
+        if (state == JFileChooser.APPROVE_OPTION) {
             try {
-                System.out.println(f.toString());
                 this.parent.getEditor().setText("");
+                complete = true;
                 try {
                     this.parent.getEditor().setText(new Scanner(f).useDelimiter("\\Z").next());
-                    complete = true;
+                    this.parent.setTitle(util.getWindowTitle(false, this.file.getName()));
                 } catch (NoSuchElementException e) {
                     System.out.println("Selected file is empty");
                 }
@@ -61,13 +60,13 @@ public class EzNoteDocument {
         } else {
             System.out.println("Unable to load file");
         }
-        this.dc.debugChange();
+        this.debugChange();
         return complete;
     }
     
     public boolean saveQ(){
         boolean saved = false;
-        if(this.dc.isFileChanged()){
+        if(this.isFileChanged()){
             boolean operation = this.saveAs();
             if(operation){
                 saved = true;
@@ -75,7 +74,7 @@ public class EzNoteDocument {
         } else {
             //manual save
         }
-        this.dc.debugChange();
+        this.debugChange();
         return saved;
     }
     
@@ -131,8 +130,107 @@ public class EzNoteDocument {
         } else {
             System.out.println("JFileChooser closed");
         }
-        this.dc.debugChange();
+        this.debugChange();
         return complete;
+    }
+
+    @Override
+    public void insertUpdate(DocumentEvent de) {
+        this.changeIndicator = true;
+        this.changeCounter++;
+        this.parent.setTitle(this.parent.getUtilities().getWindowTitle(true, this.parent.getFilename()));
+        notifyEvent("event insert");
+    }
+
+    @Override
+    public void removeUpdate(DocumentEvent de) {
+        this.changeIndicator = true;
+        this.changeCounter++;
+        notifyEvent("event remove");
+    }
+
+    @Override
+    public void changedUpdate(DocumentEvent de) {
+        this.changeIndicator = true;
+        this.changeCounter++;
+        notifyEvent("event change");
+    }
+    
+    /**
+     * Prints a string to the console (debug)
+     * Deprecated, this will be removed in some releases
+     * @param str
+     */
+    @Deprecated
+    public void notifyEvent(String str){
+        System.out.println("EzFrameDL dc > "+str);
+    }
+    
+    public void debugChange(){
+        System.out.println("EzFrameDL ch > "+this.getChangeIndicator()+" - "+this.getChangeCounter());
+    }
+
+    /**
+     * Returns the change counter of the document
+     * @return int counter change
+     */
+    public int getChangeCounter() {
+        return changeCounter;
+    }
+
+    /**
+     * Returns the change indicator of the document
+     * @return boolean change indicator
+     */
+    public boolean getChangeIndicator() {
+        return changeIndicator;
+    }
+    
+    /**
+     * Returns if the file has been changed or not
+     * @return boolean file changed ?
+     */
+    public boolean isFileChanged(){
+        return this.getChangeIndicator()==true && this.getChangeCounter()>0;
+    }
+
+    /**
+     * Returns the parent window of the listener
+     * @return parent window
+     */
+    public EzNoteFrame getParent() {
+        return parent;
+    }
+
+    /**
+     * Set the change counter of the document
+     * Deprecated: will be replaced by resetChange, this is for
+     * debug only
+     * @param changeCounter
+     */
+    @Deprecated
+    public void setChangeCounter(int changeCounter) {
+        this.changeCounter = changeCounter;
+    }
+
+    /**
+     * Set the change indicator
+     * Deprecated: will be replaced by resetChange, this is for
+     * debug only
+     * @param changeIndicator
+     */
+    @Deprecated
+    public void setChangeIndicator(boolean changeIndicator) {
+        this.changeIndicator = changeIndicator;
+    }
+    
+    /**
+     * Resets the changes censors of the document
+     */
+    public void resetIndicators(){
+        this.changeIndicator = false;
+        this.changeCounter = 0;
+        System.out.println("EzFrameDL ch > reset "+this.getChangeIndicator()+" - "+this.getChangeCounter());
     }
     
 }
